Refactor sensor classes.
Creates a sensor library providing a SerialSensor base class to handle
the common serial I/O functionality between sensors. The
uAirDataComputer class now inherits from this, and is also renamed to
the uADC class for conciseness.

The SerialSensor class has a virtual readData slot that should be
implemented by each particular sensor class to read in data to the
buffer and then parse it. It might be possible to refactor this further
to reduce code duplication, but this should work for now.

A measurementUpdate signal should be emitted when new data is parsed and
available for use.

Also includes some minor bugfixes.