Fix structure alignment for VN200 sensor.
Changes the structure alignment to 1 byte for *only* the VN200Packet
structure. Previously #pragma pack(1) was set globally, which caused
innumerable errors and segfaults. This fixes that by only setting the
packing for the struct that needs to match the VN200 data wire format.

Also changes the initialization of members to use the new C++11 style.