Request RCIN/RCOUT streams at higher rate.
Adds code to use either the REQUEST_DATA_STREAM or MESSAGE_INTERVAL
interfaces to request either the MAV_DATA_STREAM_RC_CHANNELS stream or
the RC_CHANNELS_RAW and SERVO_OUTPUT_RAW messages depending on the
interface. The data stream interface is deprecated, but the message
interval interface is not yet supported on ArduPilot. Choice of interval
and requested data rate is configured in the RC file.