Fix detection of Micro Air Data Computer.
Changes the detection algorithm for the uADC to obtain two candidate
packets if we have newlines in the data buffer, and then check that at
least one of them validates correctly.

Also fixes a bug in the serial port handling where the serial port was
not closed if a sensor was detected, but only if we failed to detect a
sensor.