Change VN-200 serial code.
Replaces the parsing code for VN-200 packets with code based on the
uADC. The effective change is that, instead of reading a single byte at
a time, the entire serial buffer is read. This simplifies the parsing
code and seems to speed up the code by an order of magnitude.