Simplify logger sensor support.
Simplifies the logger code handling detection of sensors by setting
binary flags, removing the multiple bitmask checks. Also adds a check to
prevent writing data when none is available for all of the fast sensors.
This differs from the other check in that it waits for all of the
sensors before writing.