Fix VN-200 checksum bug on amd64.
Fixes a bug in the new VN-200 CRC16 checksum implementation on (at
least) the amd64 architecture. With the previous implementation, testing
found that the checksum was always failing for valid packets because of
an integer size issue when obtaining the i-th byte of the QByteArray,
which was being converted to an int. This size issue was resolved by
static casting the byte to a quint8. This did not manifest in the old
VN-200 parsing code as the packet was stored as an array of quint8s.

Also changes the casts in the validateVN200Checksum function to C++
style static_casts instead of C-style casts.