Allow user to specify serial port baud rates.
Changes the configuration file and settings manager to support
user-defined baud rates. Currently, baud rates of 57600 and 115200 are
supported, as these are the most likely baud rates to be used and are
also supported by default as QSerialPort::BaudRate items. To set the
baud rate, use the baud_rate parameter in the RC file.

Also changes labels of debug text to be more verbose.