Change LLA from float to double in VN200Data.
Changes the data type of the latitude, longitude, and altitude messages
in the VN200Data structure from float to double so as not to lose
precision. The VN-200 sensor reports these as floats, but in the
previous version of the VN200Data class these were truncated to single
precision floating points, but then logged with the precision expected
from a double precision float.