Add option to wait for data update before logging.
Adds an option 'wait_for_update' to make the logger wait for new data
before writing to a log file. This should remove the amount of duplicate
entries in the log files while maintaining logging at the maximum rate
for each sensor.

The logic for the test is (!waitForUpdate || newData), which evaluates
to true except when waitForUpdate is true and newData is false.