Move ADC delay into analogRead loop.
Moves the delay call from the end of the loop function to inside the
for loop reading the analog values in to allow for the ADC to settle so
that we get good results from the potentiometers. This will probably
reduce the maximum sampling rate.

Tested with 2 potentiometers on an Arduino Uno with the latest version
of DFTI from master.