Add RIO support to DFTI.
Adds support for a basic Remote I/O device serial protocol. The RIO
device sends ASCII serial messages with the format
	$$$value1$value2$...$valuen$checksum\r\n
where each value is a floating point value and the checksum is a single
byte represented in hexadecimal. The checksum algorithm is a simple
bitwise XOR of each byte in the packet, non-inclusive of the checksum
byte itself.

This representation allows for the inclusion of arbitrary values,
although the intended use is direct measurement of control effector
deflections and engine RPM.

Some behavior of the DFTI logger is changed to handle the arbitrary RIO
data, which is stored as a std::vector<float>. Since the names of the
data are not necessarily known, values are logged as 'rio_value_i',
where i is an index. For this to work, the log file writing is changed
so that the first write does not occur until a data update has occured.
This /should/ in theory ensure that correct labels for the RIO CSV file
are written, although this has not been tested with multiple sensors.

This commit should resolve #12, at least on the DFTI side. The RIO
firmware needs to be updated to use the current serial protocol.