Implement UDP server to close #1.
Implements a UDP server to publish state data for other software to use
during flights for applications such as online sysid, health monitoring,
and state estimation.

A UDP server sends a StateData object with a GPS timestamp, INS attitude
(Euler angles and quaternion), angular rates, and accelerations, Air
Data System IAS, AOA, and sideslip, and RIOs for engine RPM and control
effector deflections at a user-specified rate to a user-specified
address and port.

The configuration file expects a [server] section with the following
fields: enabled (boolean), address (string, must be a valid IP address),
port (uint16_t), and rate (uint8_t, in Hertz). By default the server is
disabled.

At the moment, the server does not respect the INS wait_for_gps option
and sends data immediately.

Documentation on the StateData struct and how to use it is provided in
Doxygen.