Remove last QPointers.
Use standard pointers instead of QPointers so the Travis CI build works.
For some reason using QPointers works on later versions of Qt and the
version of Qt 5.3 on Debian, but not with the version on Travis CI.
Although we could just use QPointers and not worry about Travis CI, it's
nice to have the CI working with 5.3 so we can verify we don't break the
build on commits and for pull requests.