Use QPointers in dftitest to handle mem mgmt.
Replace standard 'dumb' pointers with Qt smart QPointers for automatic
memory management to avoid resource leaks in dftitest. Note that this
should also be done in the main dfti executable (and really throughout
the codebase) but the Qt5.3 version in the CI really doesn't seem to
like that, so we need to figure out how to make it happy.