Switch to using QPointers for DMA objects.
Replaces C and C++ style pointers with smart QPointers for objects that
are allocated dynamically to ensure proper memory management. Since
previous attempts broke the CI build (but not the native or
cross-compile builds), the qptrutil.hh header is introduced. This header
defines macros that use the QPointer Operator T*() to cast the QPointer
to a raw pointer when TRAVISCI is defined. Otherwise, the QPointer is
used directly.

This commit also bumps the CMake version to 2.2.3 in preparation for the
merge of the refactor branch.